WIP filedrop

+++ filedrop/tools/chocolateybeforemodify.ps1
@@ -1,2 +1,6 @@
 ﻿
-Stop-Process -Name "FileDrop"
+$ProcessActive = Get-Process FileDrop -ErrorAction SilentlyContinue
+if($ProcessActive -eq $null)
+{
+    Stop-Process -Name "FileDrop"
+}
